+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Lots of `RString` methods: `capacity`, `cmp`, `comparable`, `drop_bytes`,
   `dump`, `ellipsize`, `offset`, `plus`, `times`, `replace`, `scrub`,
   `shared_replace`, `split`, `update`.
+- `RRegexp::new`/`new_str`.
 
 ### Changed
 - When converting Ruby values to `RArray` (or `Vec<T>`, `[T; 1]`, or `(T,)`),